requirements.txt
asgiref==3.7.2
attrs==23.2.0
black==24.2.0
certifi==2024.2.2
charset-normalizer==2.1.1
click==8.1.7
Django==4.2.9
djangorestframework==3.14.0
drf-spectacular==0.27.1
idna==3.6
inflection==0.5.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
mypy-extensions==1.0.0
packaging==24.0
pathspec==0.12.1
platformdirs==4.2.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
referencing==0.33.0
requests==2.28.1
rpds-py==0.18.0
sqlparse==0.4.4
tomli==2.0.1
tqdm==4.66.2
typing_extensions==4.10.0
uritemplate==4.1.1
urllib3==1.26.18
